update read_gravity_field test
