+++ software/python/tests/astro/time/test_data.py
@@ -2,43 +2,8 @@ import os
 import pytest
 
 import numpy as np
-import scipy
 
-from ...conftest import custom_isclose, custom_allclose
-
-
-def load_matlab_data(file_path: str, keys: list) -> dict:
-    """Load MATLAB .mat file data and handle structures.
-
-    Args:
-        file_path (str): Path to the .mat file
-        keys (list): List of variable names to grab
-
-    Returns:
-        dict [str, np.ndarray or dict]: Dictionary of the input keys and their
-                                        associated MATLAB data as numpy arrays or dicts
-    """
-
-    def unpack_structure(struct):
-        """Recursively unpack MATLAB structure arrays into dictionaries."""
-        if isinstance(struct, np.ndarray) and struct.dtype.names:
-            return {name: struct[name] for name in struct.dtype.names}
-        return struct
-
-    # Load the .mat file
-    data = scipy.io.loadmat(file_path, struct_as_record=False, squeeze_me=True)
-
-    # Process keys and unpack structures
-    result = {}
-    for key in keys:
-        if key in data:
-            value = data[key]
-            if isinstance(value, np.ndarray) and value.dtype.names:
-                # If the key is a structure, unpack its fields
-                result[key] = unpack_structure(value)
-            else:
-                result[key] = value
-    return result
+from ...conftest import custom_isclose, custom_allclose, load_matlab_data
 
 
 @pytest.fixture()
